+++ CMakeLists.txt
@@ -13,11 +13,11 @@ endif()
 add_library(encoder OBJECT 
     src/Encoder.h
     src/Encoder.cpp
-  "src/Compression/shannon_fano.h")
+  "src/Compression/shannon_fano.h" "src/Compression/compression.h")
 
 add_library(decoder OBJECT 
    src/Decoder.h
-    src/Decoder.cpp "src/Compression/shannon_fano.h")
+    src/Decoder.cpp "src/Compression/shannon_fano.h" "src/Compression/compression.h")
 
 
 add_executable(${CMAKE_PROJECT_NAME}
@@ -26,13 +26,13 @@ add_executable(${CMAKE_PROJECT_NAME}
     src/h4z0t.h
       
  "src/FileRegister.h"
-"src/File.h" "src/Compression/shannon_fano.h")
+"src/File.h" "src/Compression/shannon_fano.h" "src/Compression/compression.h")
 
 set( CMAKE_PROJECT_NAME_TEST "${CMAKE_PROJECT_NAME}_Test")
 
 add_executable(${CMAKE_PROJECT_NAME_TEST}
     src/test_compression/sf.cpp
-    )
+     "src/Compression/compression.h")
 
  set_target_properties(${CMAKE_PROJECT_NAME_TEST} PROPERTIES
         CXX_STANDARD 17
